requirement.txt
tensorflow-gpu==1.12.0
sklearn==0.0
numpy==1.14.0+mkl
gym==0.10.9
matplotlib==3.0.1
pandas==0.24.2